# docs/swagger.yaml
definitions:
...
  handlers.GetUserHandlerData:
    properties:
      categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      issues:
        items:
          $ref: '#/definitions/models.Issue'
        type: array
      projects:
        items:
          $ref: '#/definitions/models.Project'
        type: array
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      trackers:
        items:
          $ref: '#/definitions/models.Tracker'
        type: array
      user:
        $ref: '#/definitions/models.User'
    type: object
  handlers.GetUsersHandlerData:
    properties:
      count:
        type: integer
      projects:
        items:
          $ref: '#/definitions/models.Project'
        type: array
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      trackers:
        items:
          $ref: '#/definitions/models.Tracker'
        type: array
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
      users_roles:
        items:
          $ref: '#/definitions/models.UserRole'
        type: array
    type: object
  models.Category:
    properties:
      assigned_to_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      project_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.Issue:
    properties:
      assigned_to_id:
        type: integer
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      project_id:
        type: integer
      status:
        type: string
      subject:
        type: string
      tracker_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.Project:
    properties:
      created_on:
        type: string
      description:
        type: string
      id:
        type: integer
      identifier:
        type: string
      name:
        type: string
      parent_id:
        description: Usamos *int para manejar valores NULL
        type: integer
      updated_on:
        type: string
    type: object
  models.Role:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  models.Tracker:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      password_hash:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.UserRole:
    properties:
      role_id:
        type: integer
      user_id:
        type: integer
    type: object
info:
  contact: {}
  description: API para gestionar proyectos, usuarios, roles, categorías y problemas
    al estilo de Redmine.
  title: Go Redmine-ish API
  version: "1.0"
paths:
...
  /user:
    post:
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
  /user/{id}:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.GetUserHandlerData'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
    put:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
  /users:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.GetUsersHandlerData'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: 'Ingresa el token en formato: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"